Add MINIPAL_ATTR_FORMAT_PRINTF to remaining printf-like wrappers
Annotate log_init_error_to_host, LogErrorToHost, _MulticoreJitTrace,
FormatString, CrashInfo::Trace, CrashInfo::TraceVerbose, trace_printf,
trace_verbose_printf, printf_status, and printf_error with
MINIPAL_ATTR_FORMAT_PRINTF so the compiler checks format strings.

Fix format mismatches found by the new annotations:
- crashinfomac.cpp: cast pages_resident to unsigned long long for %llx;
  use %zd/%lu for size_t/ssize_t/vm_size_t args
- createdumpmain.cpp: use PRIu64 for uint64_t g_ticksPerMS and elapsed
- createdumpunix.cpp: use %lu cast for PAGE_SIZE (vm_size_t)
- dumpwritermacho.cpp: use %zu for size_t; use PRIu64 for uint64_t
- interface.cpp: use %.3f for float ratio values in log_init_error_to_host

Also fix pre-existing broken string literals in collect.cpp and
dynamic_heap_count.cpp where trailing commas were inside the format
string instead of separating it from arguments.

+++ src/coreclr/debug/createdump/createdumpunix.cpp
@@ -22,7 +22,7 @@ CreateDump(const CreateDumpOptions& options)
 #if defined(__arm__) || defined(__aarch64__) || defined(__loongarch64) || defined(__riscv)
     g_pageSize = sysconf(_SC_PAGESIZE);
 #endif
-    TRACE("PAGE_SIZE %d\n", PAGE_SIZE);
+    TRACE("PAGE_SIZE %lu\n", (unsigned long)PAGE_SIZE);
 
     if (options.CrashReport && (options.AppModel == AppModelType::SingleFile || options.AppModel == AppModelType::NativeAOT))
     {
